feat: Add CustomModal component
Add a new component called CustomModal that displays a modal overlay with a message and a close button. This component is used to show alerts or notifications to the user.

The CustomModal component accepts three props:
- visible: a boolean indicating whether the modal is visible or not
- message: a string representing the message to be displayed in the modal
- onClose: a callback function to be called when the close button is pressed

The modal is displayed with a fade animation and has a semi-transparent overlay. The modal container has a fixed width of 300 pixels and a padding of 20 pixels. The background color is white with a border radius of 10 pixels. The message text is centered and has a font size of 16 pixels. The close button has a background color of green (#23A331) and white text color.

This commit adds the CustomModal component to the project.